backend/pyproject.toml
[project]
name = "ag-ui-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.5.1",
    "copilotkit>=0.1.78",
    "fastapi>=0.115.14",
    "google-adk>=1.20.0",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
